requirements.txt
asn1crypto==0.24.0
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
cryptography==2.4.1
Django==2.1.3
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.9.0
idna==2.7
Markdown==3.0.1
Pillow==5.3.0
psycopg2==2.7.7
pycparser==2.19
pytz==2018.7
qiniu==7.2.2
requests==2.21.0
six==1.11.0
urllib3==1.24.1
